pyproject.toml
[tool.poetry]
name = "firepinnew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
